requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
gnews>=0.3.7
tavily-python>=0.3.0
youtube-search>=2.1.0
wikipedia>=1.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyperclip>=1.8.2
